requirements.txt
fastapi
uvicorn[standard]
ollama
requests
python-dotenv
cachetools
